Refactor Action Score to be double

+++ src/main/java/org/jgbakke/jlearning/ActionScore.java
@@ -2,14 +2,14 @@ package org.jgbakke.jlearning;
 
 public class ActionScore {
     public final Action action;
-    public final int score;
+    public final double score;
 
-    public ActionScore(Action action, int score) {
+    public ActionScore(Action action, double score) {
         this.action = action;
         this.score = score;
     }
 
-    public ActionScore(int actionID, int score){
+    public ActionScore(int actionID, double score){
         this(ActionContainer.getInstance().getActionById(actionID), score);
     }
